pyproject.toml
[tool.poetry]
name = "sv-test-scripts"
version = "0.1.0"
description = ""
authors = ["JoseVL92 <jovalab92@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.28.1"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.3"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
